Huge Update
Added a timer to the RO app, which now reflects the time entered from the first screen. Also removed Pandas, NumPy and io from the imports to hopefully make the file smaller and the program run a little faster.
Fixed more issues with error handling and cleaned up the code a little to be more readable.