Readability Update
Reorganized code so that similar functions are now grouped together. The groups are organized as: "Time Conversions/Data Processing", "Competition Logic", "Button/Action Functions", "UI Management" and "Timer"

Added the ability to create a timer without having to first set up a round with a start time/list of climbers.