Holy Organization Batman
Reorganization of files so the project can be more readable. Working towards aligning project with the natural style that GitHub projects seem to use.
Main .py files can be found in the src folder (as well as necessary sound files).
notebooks contains the notebooks used for experimentation and testing.
assets contains pictures for the README file.